Extract new services/documents.js file
